refactor: extract path matching logic to helper class
